requirements.txt
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
Pillow==8.4.0
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
